+++ src/main/java/com/godeltech/facade/impl/AirportFacadeImpl.java
@@ -1,9 +1,12 @@
 package com.godeltech.facade.impl;
 
+import com.godeltech.component.LocalMessageSource;
+import com.godeltech.exception.AssignmentException;
 import com.godeltech.facade.AirportFacade;
 import com.godeltech.mapper.AirportMapper;
 import com.godeltech.persistence.model.Airport;
 import com.godeltech.service.AirportService;
+import com.godeltech.service.FlightService;
 import com.godeltech.web.dto.request.AirportRequestDto;
 import com.godeltech.web.dto.response.AirportResponseDto;
 import lombok.RequiredArgsConstructor;
@@ -11,11 +14,14 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Component
 @Slf4j
 @RequiredArgsConstructor
 public class AirportFacadeImpl implements AirportFacade {
+    private final FlightService flightService;
+    private final LocalMessageSource messageSource;
     private final AirportService airportService;
     private final AirportMapper airportMapper;
 
@@ -41,7 +47,7 @@ public class AirportFacadeImpl implements AirportFacade {
     @Override
     public AirportResponseDto update(final Long id, final AirportRequestDto airportRequestDto) {
         log.debug("Update airport with id:{}", id);
-        Airport airport = airportService.findById(id);
+        final Airport airport = airportService.findById(id);
         airportMapper.updateEntity(airport, airportRequestDto);
         return airportMapper.toAirportResponseDto(airportService.update(airport));
     }
@@ -49,6 +55,14 @@ public class AirportFacadeImpl implements AirportFacade {
     @Override
     public void deleteById(final Long id) {
         log.debug("Delete airport with id:{}", id);
-        airportService.deleteById(id);
+        final boolean isAssignedToFlight = flightService.findAll().stream()
+                .map(flight -> flight.getDepartureAirport().getId())
+                .collect(Collectors.toList())
+                .contains(id);
+        if (isAssignedToFlight) {
+            throw new AssignmentException(messageSource.getMessage("error.record.isAssignment", new Object[]{}));
+        } else {
+            airportService.deleteById(id);
+        }
     }
 }
