+++ src/main/java/com/godeltech/facade/impl/CategoryFacadeImpl.java
@@ -1,9 +1,13 @@
 package com.godeltech.facade.impl;
 
+import com.godeltech.component.LocalMessageSource;
+import com.godeltech.exception.AssignmentException;
 import com.godeltech.facade.CategoryFacade;
 import com.godeltech.mapper.CategoryMapper;
 import com.godeltech.persistence.model.Category;
+import com.godeltech.service.AirplaneService;
 import com.godeltech.service.CategoryService;
+import com.godeltech.service.FlightService;
 import com.godeltech.web.dto.request.CategoryRequestDto;
 import com.godeltech.web.dto.response.CategoryResponseDto;
 import lombok.RequiredArgsConstructor;
@@ -11,11 +15,14 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Component
 @Slf4j
 @RequiredArgsConstructor
 public class CategoryFacadeImpl implements CategoryFacade {
+    private final AirplaneService airplaneService;
+    private final LocalMessageSource messageSource;
     private final CategoryService categoryService;
     private final CategoryMapper categoryMapper;
 
@@ -40,7 +47,7 @@ public class CategoryFacadeImpl implements CategoryFacade {
     @Override
     public CategoryResponseDto update(final Long id,final CategoryRequestDto categoryRequestDto) {
         log.debug("Update category with id:{}", id);
-        Category category = categoryService.findById(id);
+        final Category category = categoryService.findById(id);
         categoryMapper.updateEntity(category,categoryRequestDto);
         return categoryMapper.toCategoryResponseDto(categoryService.update(category));
     }
@@ -48,6 +55,14 @@ public class CategoryFacadeImpl implements CategoryFacade {
     @Override
     public void deleteById(final Long id) {
         log.debug("Delete category with id:{}", id);
-        categoryService.deleteById(id);
+        final boolean isAssignedToFlight = airplaneService.findAll().stream()
+                .map(flight -> flight.getCategory().getId())
+                .collect(Collectors.toList())
+                .contains(id);
+        if (isAssignedToFlight) {
+            throw new AssignmentException(messageSource.getMessage("error.record.isAssignment", new Object[]{}));
+        } else {
+            categoryService.deleteById(id);
+        }
     }
 }
