+++ src/main/java/com/godeltech/facade/impl/SecondPilotFacadeImpl.java
@@ -1,8 +1,11 @@
 package com.godeltech.facade.impl;
 
+import com.godeltech.component.LocalMessageSource;
+import com.godeltech.exception.AssignmentException;
 import com.godeltech.facade.SecondPilotFacade;
 import com.godeltech.mapper.SecondPilotMapper;
 import com.godeltech.persistence.model.SecondPilot;
+import com.godeltech.service.FlightService;
 import com.godeltech.service.SecondPilotService;
 import com.godeltech.web.dto.request.SecondPilotRequestDto;
 import com.godeltech.web.dto.response.SecondPilotResponseDto;
@@ -11,13 +14,16 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Component
 @Slf4j
 @RequiredArgsConstructor
 public class SecondPilotFacadeImpl implements SecondPilotFacade {
+    private final FlightService flightService;
     private final SecondPilotService secondPilotService;
     private final SecondPilotMapper secondPilotMapper;
+    private final LocalMessageSource messageSource;
 
     @Override
     public SecondPilotResponseDto findById(final Long id) {
@@ -40,7 +46,7 @@ public class SecondPilotFacadeImpl implements SecondPilotFacade {
     @Override
     public SecondPilotResponseDto update(final Long id, final SecondPilotRequestDto secondPilotRequestDto) {
         log.debug("Update second pilot with id:{}", id);
-        SecondPilot secondPilot = secondPilotService.findById(id);
+        final SecondPilot secondPilot = secondPilotService.findById(id);
         secondPilotMapper.updateEntity(secondPilot, secondPilotRequestDto);
         return secondPilotMapper.toSecondPilotResponseDto(secondPilotService.update(secondPilot));
     }
@@ -48,6 +54,15 @@ public class SecondPilotFacadeImpl implements SecondPilotFacade {
     @Override
     public void deleteById(final Long id) {
         log.debug("Delete second pilot with id:{}", id);
-        secondPilotService.deleteById(id);
+        final boolean isAssignedToFlight = flightService.findAll().stream()
+                .map(flight -> flight.getSecondPilot().getId())
+                .collect(Collectors.toList())
+                .contains(id);
+        if (isAssignedToFlight) {
+            throw new AssignmentException(messageSource.getMessage("error.record.isAssignment", new Object[]{}));
+        } else {
+            secondPilotService.deleteById(id);
+        }
+
     }
 }
